ASC-54 Changed location of some retrieval parameters in 'derived_parameters' and corrected incorrect param names
